refactor(EventDetails): convert EditEventPage to EventDetailPage, enhancing event display and functionality
